IGNITE-23114 REVERT subclass for BinaryReaderExImpl
