+++ modules/core/src/main/java/org/apache/ignite/internal/processors/odbc/jdbc/JdbcBinaryTypeNameGetRequest.java
@@ -21,7 +21,7 @@ package org.apache.ignite.internal.processors.odbc.jdbc;
 import org.apache.ignite.binary.BinaryObjectException;
 import org.apache.ignite.internal.MarshallerPlatformIds;
 import org.apache.ignite.internal.binary.BinaryReaderExImpl;
-import org.apache.ignite.internal.processors.odbc.SqlBinaryWriter;
+import org.apache.ignite.internal.binary.BinaryWriterExImpl;
 import org.apache.ignite.internal.util.typedef.internal.S;
 
 /**
@@ -69,7 +69,7 @@ public class JdbcBinaryTypeNameGetRequest extends JdbcRequest {
     }
 
     /** {@inheritDoc} */
-    @Override public void writeBinary(SqlBinaryWriter writer, JdbcProtocolContext protoCtx)
+    @Override public void writeBinary(BinaryWriterExImpl writer, JdbcProtocolContext protoCtx)
         throws BinaryObjectException {
         super.writeBinary(writer, protoCtx);
