+++ modules/core/src/main/java/org/apache/ignite/internal/processors/odbc/jdbc/JdbcBinaryTypeNameGetResult.java
@@ -20,7 +20,7 @@ package org.apache.ignite.internal.processors.odbc.jdbc;
 
 import org.apache.ignite.binary.BinaryObjectException;
 import org.apache.ignite.internal.binary.BinaryReaderExImpl;
-import org.apache.ignite.internal.processors.odbc.SqlBinaryWriter;
+import org.apache.ignite.internal.binary.BinaryWriterExImpl;
 import org.apache.ignite.internal.util.typedef.internal.S;
 
 /**
@@ -68,7 +68,7 @@ public class JdbcBinaryTypeNameGetResult extends JdbcResult {
     }
 
     /** {@inheritDoc} */
-    @Override public void writeBinary(SqlBinaryWriter writer, JdbcProtocolContext protoCtx)
+    @Override public void writeBinary(BinaryWriterExImpl writer, JdbcProtocolContext protoCtx)
         throws BinaryObjectException {
         super.writeBinary(writer, protoCtx);
