+++ modules/core/src/main/java/org/apache/ignite/internal/processors/odbc/jdbc/JdbcBinaryTypeNamePutRequest.java
@@ -21,7 +21,7 @@ package org.apache.ignite.internal.processors.odbc.jdbc;
 import org.apache.ignite.binary.BinaryObjectException;
 import org.apache.ignite.internal.MarshallerPlatformIds;
 import org.apache.ignite.internal.binary.BinaryReaderExImpl;
-import org.apache.ignite.internal.processors.odbc.SqlBinaryWriter;
+import org.apache.ignite.internal.binary.BinaryWriterExImpl;
 import org.apache.ignite.internal.util.typedef.internal.S;
 
 /**
@@ -83,7 +83,7 @@ public class JdbcBinaryTypeNamePutRequest extends JdbcRequest {
     }
 
     /** {@inheritDoc} */
-    @Override public void writeBinary(SqlBinaryWriter writer, JdbcProtocolContext protoCtx)
+    @Override public void writeBinary(BinaryWriterExImpl writer, JdbcProtocolContext protoCtx)
         throws BinaryObjectException {
         super.writeBinary(writer, protoCtx);
