+++ modules/core/src/main/java/org/apache/ignite/internal/processors/odbc/jdbc/JdbcUpdateBinarySchemaResult.java
@@ -19,7 +19,7 @@ package org.apache.ignite.internal.processors.odbc.jdbc;
 
 import org.apache.ignite.binary.BinaryObjectException;
 import org.apache.ignite.internal.binary.BinaryReaderExImpl;
-import org.apache.ignite.internal.processors.odbc.SqlBinaryWriter;
+import org.apache.ignite.internal.binary.BinaryWriterExImpl;
 import org.apache.ignite.internal.util.typedef.internal.S;
 
 /**
@@ -67,7 +67,7 @@ public class JdbcUpdateBinarySchemaResult extends JdbcResult {
     }
 
     /** {@inheritDoc} */
-    @Override public void writeBinary(SqlBinaryWriter writer, JdbcProtocolContext protoCtx) throws BinaryObjectException {
+    @Override public void writeBinary(BinaryWriterExImpl writer, JdbcProtocolContext protoCtx) throws BinaryObjectException {
         super.writeBinary(writer, protoCtx);
 
         writer.writeLong(reqId);
